fix: support parsing IPv6 addresses with zone indexes in browsers
IPv6 addresses like `fe80::8cb1:25ff:fec5:28e3%llw0` are valid but
contain a zone index.  This is parsed correctly by the `net` module
on node but not by the browser polyfill here.

The fix here is just to strip the `%` character and anyything that
follows it from the input string.

Adds a test for this, and also adds support for running the tests
in Chrome in dev and CI.

Fixes #9